requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
streamlit
tiktoken
pypdf

